requirements.txt
streamlit==1.16.0
PyGithub==1.55